+++ packages/cms/cmsfilter/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fixed `undefined` errors when checking range validity in single range props.
 - Fixed value checking when multiple non checkbox/radio elements exist in a filter key.
 - Fixed `fs-cmsfilter-range` filtering issue when the `from` value was not specified.
+- Fixed `fs-cmsfilter-range` values not being correctly set from the URL query params on page load.
+
 - Improved tags rendering when the filter is a `range`. Now it outputs a single tag with a `[FROM_VALUE, TO_VALUE]` format, replacing the `undefined` values with a double dash `--`.
 - Exposed the `CMSFilters` instances at `window.fsAttributes.cms.filtersInstances`.